+++ services/gateway/pkg/product/contracts/update_product_request.go
@@ -7,7 +7,7 @@ type UpdateProductRequest struct {
 	Name        string  `json:"name" validate:"required,max=50"`
 	Description *string `json:"description" validate:"omitempty,max=50"`
 	Sku         *string `json:"sku" validate:"omitempty,max=12"`
-	Quantity    uint64  `json:"quantity" validate:"required"`
+	Quantity    uint32  `json:"quantity" validate:"required"`
 	Price       uint64  `json:"price" validate:"required"`
 }
